-Refactored info item and local item click gestures into the same OnClickGesture.
