--- app/src/main/java/org/schabi/newpipe/info_list/OnInfoItemGesture.java
@@ -1,12 +1,0 @@
-package org.schabi.newpipe.info_list;
-
-import org.schabi.newpipe.extractor.InfoItem;
-
-public abstract class OnInfoItemGesture<T extends InfoItem> {
-
-    public abstract void selected(T selectedItem);
-
-    public void held(T selectedItem) {
-        // Optional gesture
-    }
-}
